+++ lib/specifications/projects/repository-types/git.js
@@ -9,7 +9,7 @@ var _ = require('underscore');
 var fs = require('q-io/fs'); // https://github.com/kriskowal/q-io
 var util = require('util');
 
-var appConfig = require('../../configuration/app-config').get();
+var appConfig = require('../../../configuration/app-config').get();
 
 
 var ignoreSSLErrors = {
